User flows working. Unit and integration tests passing.

+++ app.py
@@ -4,6 +4,7 @@ from flask_debugtoolbar import DebugToolbarExtension
 from models import connect_db, db, Book, User, UserBook, Tag, UserTag, UserBookTag
 from forms import UserForm
 from utils import lookup_isbn_open_library, map_response_to_book
+from sqlalchemy.exc import IntegrityError
 
 app = Flask(__name__)
 
@@ -41,41 +42,42 @@ def do_login(user):
 
 
 def do_logout():
-    """Lof the user out."""
+    """Log the user out."""
 
     if CURR_USER_KEY in session:
         del session[CURR_USER_KEY]
-
-
-@app.route('/')
-def home():
-    """Redirect the user to the appropriate route based on whether or not they are logged in."""
-
-    if g.user:
-        return redirect(f'/users/{g.user.id}/books')
-
-    else:
-        return render_template('home-anon.html')
 
 
 @app.route('/signup', methods=['GET', 'POST'])
 def signup():
     """Sign up a user."""
 
+    # import pdb
+    # pdb.set_trace()
+
+    # if there is a user logged ask them to logout before creating a new user
+    if g.user:
+        flash("Please logout before creating a new account", "danger")
+        return redirect('/')
+
     form = UserForm()
 
     if form.validate_on_submit():
-        user = User.signup(
-            username=form.username.data,
-            password=form.password.data
-        )
-        db.session.commit()
+        try:
+            user = User.signup(
+                username=form.username.data,
+                password=form.password.data
+            )
+            db.session.commit()
+        except IntegrityError:
+            flash("Username already taken", "danger")
+            return render_template('signup.html', form=form)
 
         do_login(user)
 
         return redirect('/')
 
-    return render_template('signup.html', form=form, user=g.user)
+    return render_template('signup.html', form=form)
 
 
 @app.route('/login', methods=['GET', 'POST'])
@@ -106,17 +108,28 @@ def logout():
     return redirect('/')
 
 
-@app.route('/users/<int:user_id>')
-def user_home(user_id):
-    """Displays the user's main page."""
+@app.route('/')
+def home():
+    """Redirect the user to the appropriate route based on whether or not they are logged in."""
 
     if g.user:
-        if g.user.id != user_id:
-            flash("You are not authorized.", "danger")
-    else:
-        flash("You are not authorized.", "danger")
+        return redirect(f'/users/{g.user.id}/books')
 
-    return redirect('/')
+    else:
+        return render_template('home-anon.html')
+
+
+# @app.route('/users/<int:user_id>')
+# def user_home(user_id):
+#     """Displays the user's main page."""
+#
+#     if g.user:
+#         if g.user.id != user_id:
+#             flash("You are not authorized.", "danger")
+#     else:
+#         flash("You are not authorized.", "danger")
+#
+#     return redirect('/')
 
 
 @app.route('/books/search', methods=['POST'])
@@ -146,7 +159,7 @@ def search_isbn():
     return redirect(f'/books/{book.id}')
 
 
-@app.route('/books/<int:book_id>', methods=['GET', 'POST'])
+@app.route('/books/<int:book_id>', methods=['GET'])     # removed post
 def book_detail(book_id):
     """Show the searched books information."""
 
@@ -168,7 +181,7 @@ def user_books(user_id):
         return redirect('/')
 
     if g.user.id != user_id:
-        flash('You are not authorized!', 'danger')
+        flash('You are not authorized.', 'danger')
         return redirect('/')
 
     books = db.session.query(Book).join(UserBook).filter(UserBook.user_id == user_id).all()
@@ -185,7 +198,7 @@ def user_book_detail(user_id, book_id):
         return redirect('/')
 
     if g.user.id != user_id:
-        flash('You are not authorized!', 'danger')
+        flash('You are not authorized.', 'danger')
         return redirect('/')
 
     book = Book.query.filter_by(id=book_id).first()
@@ -208,12 +221,13 @@ def user_book_detail(user_id, book_id):
         )
 
     if request.method == 'POST':
-        book_user = UserBook(
-            user_id=user_id,
-            book_id=book_id
-        )
-        db.session.add(book_user)
-        db.session.commit()
+        if book_id not in [book.id for book in g.user.books]:
+            book_user = UserBook(
+                user_id=user_id,
+                book_id=book_id
+            )
+            db.session.add(book_user)
+            db.session.commit()
 
         return redirect(f'/users/{user_id}/books/{book_id}')
 
@@ -222,8 +236,12 @@ def user_book_detail(user_id, book_id):
 def delete_user_book(user_id, book_id):
     """Remove a book from a user's collection."""
 
+    if not g.user:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
     if g.user.id != user_id:
-        flash('You are not authorized!', 'danger')
+        flash('You are not authorized.', 'danger')
         return redirect('/')
 
     if book_id not in [book.id for book in g.user.books]:
@@ -232,6 +250,11 @@ def delete_user_book(user_id, book_id):
 
     user_book = db.session.query(UserBook).filter(UserBook.user_id == user_id, UserBook.book_id == book_id).first()
     db.session.delete(user_book)
+    book_tags = db.session.query(UserBookTag)\
+        .filter(UserBookTag.user_id == user_id, UserBookTag.book_id == book_id)\
+        .all()
+    for book in book_tags:
+        db.session.delete(book)
     db.session.commit()
 
     return redirect(f'/users/{user_id}/books')
@@ -244,36 +267,38 @@ def user_tags(user_id):
     POST: Create a new user tag.
     """
 
+    if not g.user:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
     if g.user.id != user_id:
-        flash('You are not authorized!', 'danger')
+        flash('You are not authorized.', 'danger')
         return redirect('/')
 
     return render_template('user-tags.html', user=g.user)
 
 
-@app.route('/users/<int:user_id>/tag', methods=['GET', 'POST'])
+@app.route('/users/<int:user_id>/tag', methods=['POST'])
 def add_user_tag(user_id):
     """
-    GET: Show the add user tag form.
     POST: Create a user tag.
     """
 
-    if g.user.id != user_id:
-        flash('You are not authorized!', 'danger')
+    if not g.user:
+        flash('You are not authorized.', 'danger')
         return redirect('/')
 
-    if request.method == 'GET':
-        return render_template('add-user-tag.html', user=g.user)
+    if g.user.id != user_id:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
 
     if request.method == 'POST':
         tag_name = request.form.get('tag')
         tag = Tag.query.filter_by(name=tag_name).first()
         if not tag:
-            new_tag = Tag(name=tag_name)
-            db.session.add(new_tag)
+            tag = Tag(name=tag_name)
+            db.session.add(tag)
             db.session.commit()
-
-        tag = new_tag if new_tag else tag
 
         user_tag = UserTag(
             user_id=g.user.id,
@@ -285,20 +310,56 @@ def add_user_tag(user_id):
         return redirect(f'/users/{g.user.id}/tags')
 
 
-@app.route('/users/<int:user_id>/books/<int:book_id>/tags/<int:tag_id>', methods=['POST'])
+@app.route('/users/<int:user_id>/tag/<int:tag_id>/delete', methods=['POST'])
+def delete_user_tag(user_id, tag_id):
+    """
+    Remove a tag from the users' collection.
+    Remove the tag from any books in the user's collection.
+    """
+
+    if not g.user:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    if g.user.id != user_id:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    if tag_id not in [tag.id for tag in g.user.tags]:
+        flash('Tag not found!', 'danger')
+        return redirect('/')
+
+    user_tag = db.session.query(UserTag).filter(UserTag.user_id == user_id, UserTag.tag_id == tag_id).first()
+    user_book_tags = db.session.query(UserBookTag)\
+        .filter(UserBookTag.user_id == user_id, UserBookTag.tag_id == tag_id)\
+        .all()
+    db.session.delete(user_tag)
+    for user_book_tag in user_book_tags:
+        db.session.delete(user_book_tag)
+    db.session.commit()
+
+    return redirect(f'/users/{user_id}/tags')
+
+
+@app.route('/users/<int:user_id>/books/<int:book_id>/tag/<int:tag_id>', methods=['POST'])
 def add_book_tag(user_id, book_id, tag_id):
     """Add a tag to a book in a user's collection."""
-    import pdb
-    pdb.set_trace()
-    user_books = db.session.query(Book).join(UserBook).filter(UserBook.user_id == user_id).all()
-    if book_id not in [book.id for book in user_books]:
-        flash('Book not found!', 'danger')
-        return redirect(request.referrer)
 
-    user_tags = db.session.query(Tag).join(UserTag).filter(UserTag.user_id == user_id).all()
-    if tag_id not in [tag.id for tag in user_tags]:
+    if not g.user:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    if g.user.id != user_id:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    if book_id not in [book.id for book in g.user.books]:
+        flash('Book not found!', 'danger')
+        return redirect('/')
+
+    if tag_id not in [tag.id for tag in g.user.tags]:
         flash('Tag not found!', 'danger')
-        return redirect(request.referrer)
+        return redirect(f'/users/{user_id}/books/{book_id}')
 
     book_tag = UserBookTag(
         user_id=user_id,
@@ -308,37 +369,56 @@ def add_book_tag(user_id, book_id, tag_id):
     db.session.add(book_tag)
     db.session.commit()
 
-    return redirect(request.referrer)
+    return redirect(f'/users/{user_id}/books/{book_id}')
 
 
-@app.route('/users/<int:user_id>/books/<int:book_id>/tags/<int:tag_id>/delete', methods=['POST'])
+@app.route('/users/<int:user_id>/books/<int:book_id>/tag/<int:tag_id>/delete', methods=['POST'])
 def delete_book_tag(user_id, book_id, tag_id):
     """Remove a tag from a book in the user's collection."""
 
-    user_books = db.session.query(Book).join(UserBook).filter(UserBook.user_id == user_id).all()
-    if book_id not in [book.id for book in user_books]:
-        flash('Book not found!', 'danger')
-        return redirect(request.referrer)
+    if not g.user:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
 
-    user_tags = db.session.query(Tag).join(UserTag).filter(UserTag.user_id == user_id).all()
-    if tag_id not in [tag.id for tag in user_tags]:
+    if g.user.id != user_id:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    if book_id not in [book.id for book in g.user.books]:
+        flash('Book not found!', 'danger')
+        return redirect('/')
+
+    if tag_id not in [tag.id for tag in g.user.tags]:
         flash('Tag not found!', 'danger')
-        return redirect(request.referrer)
+        return redirect(f'/users/{user_id}/books/{book_id}')
 
     user_book_tag = UserBookTag.query.filter_by(user_id=user_id, book_id=book_id, tag_id=tag_id).first()
     db.session.delete(user_book_tag)
     db.session.commit()
 
-    return redirect(request.referrer)
+    return redirect(f'/users/{user_id}/books/{book_id}')
 
 
 @app.route('/users/<int:user_id>/tags/<int:tag_id>')
 def show_user_book_by_tag(user_id, tag_id):
     """Show all the books in the user's collection with the specified tag."""
 
+    if not g.user:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    if g.user.id != user_id:
+        flash('You are not authorized.', 'danger')
+        return redirect('/')
+
+    tag = Tag.query.filter_by(id=tag_id).first()
+    if tag not in g.user.tags:
+        flash('Tag not found!', 'danger')
+        return redirect('/')
+
     books = db.session.query(Book)\
         .join(UserBookTag)\
         .filter(UserBookTag.user_id == user_id, UserBookTag.tag_id == tag_id)\
         .all()
 
-    return render_template('user-books.html', user=g.user, books=books)
+    return render_template('user-books.html', user=g.user, books=books, tag=tag)
